refactor:Add dataModificationForMovies file
